backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai
langchain-experimental==0.0.58
httpx
pydantic
tabulate
pandas
numpy<2.0.0